+++ craft_providers/lxd/launcher.py
@@ -26,6 +26,7 @@ import threading
 import time
 from datetime import datetime, timedelta, timezone
 from pathlib import Path
+from typing import TYPE_CHECKING
 
 from craft_providers import Base, ProviderError, bases
 from craft_providers.errors import details_from_called_process_error
@@ -35,6 +36,11 @@ from .lxc import LXC
 from .lxd_instance import LXDInstance
 from .lxd_instance_status import LXDInstanceState, ProviderInstanceStatus
 from .project import create_with_default_profile
+
+if TYPE_CHECKING:
+    from collections.abc import Callable
+
+    from craft_providers import Executor
 
 logger = logging.getLogger(__name__)
 
@@ -95,6 +101,7 @@ def _create_instance(  # noqa: PLR0913
     project: str,
     remote: str,
     lxc: LXC,
+    prepare_instance: "Callable[[Executor], None] | None" = None,
 ) -> None:
     """Launch and setup an instance from an image.
 
@@ -115,6 +122,8 @@ def _create_instance(  # noqa: PLR0913
     :param gid: The group id to be mapped, if ``map_user_uid`` is enabled.
     :param project: LXD project to create instance in.
     :param remote: LXD remote to create instance on.
+    :param prepare_instance: A callback to perform early instance configuration
+    before the base image setup.
     :param lxc: LXC client.
     """
     logger.info("Creating new instance from remote")
@@ -146,6 +155,9 @@ def _create_instance(  # noqa: PLR0913
             )
             config_timer = InstanceTimer(base_instance)
             config_timer.start()
+
+            if prepare_instance:
+                prepare_instance(base_instance)
 
             # The base configuration shouldn't mount cache directories because if
             # they get deleted, copying the base instance will fail.
@@ -204,6 +216,10 @@ def _create_instance(  # noqa: PLR0913
             )
             config_timer = InstanceTimer(instance)
             config_timer.start()
+
+            if prepare_instance:
+                prepare_instance(instance)
+
             base_configuration.setup(executor=instance)
             _set_timezone(instance, project, remote, instance.lxc)
             instance.config_set(
@@ -686,6 +702,7 @@ def launch(  # noqa: PLR0913
     remote: str = "local",
     lxc: LXC = LXC(),  # noqa: B008
     expiration: timedelta = timedelta(days=90),
+    prepare_instance: "Callable[[Executor], None] | None" = None,
 ) -> LXDInstance:
     """Create, start, and configure an instance.
 
@@ -722,6 +739,8 @@ def launch(  # noqa: PLR0913
     :param remote: LXD remote to create instance on.
     :param lxc: LXC client.
     :param expiration: How long a base instance will be valid from its creation date.
+    :param prepare_instance: A callback to perform early instance configuration
+    before the base image setup.
 
     :returns: LXD instance.
 
@@ -782,6 +801,7 @@ def launch(  # noqa: PLR0913
             project=project,
             remote=remote,
             lxc=lxc,
+            prepare_instance=prepare_instance,
         )
         return instance
 
@@ -828,6 +848,7 @@ def launch(  # noqa: PLR0913
             project=project,
             remote=remote,
             lxc=lxc,
+            prepare_instance=prepare_instance,
         )
         return instance
 
@@ -852,6 +873,7 @@ def launch(  # noqa: PLR0913
             project=project,
             remote=remote,
             lxc=lxc,
+            prepare_instance=prepare_instance,
         )
         return instance