+++ craft_providers/multipass/_launch.py
@@ -18,9 +18,15 @@
 """Multipass Provider."""
 
 import logging
+from typing import TYPE_CHECKING
 
 from craft_providers import Base, bases
 from craft_providers.multipass.multipass_instance import MultipassInstance
+
+if TYPE_CHECKING:
+    from collections.abc import Callable
+
+    from craft_providers import Executor
 
 logger = logging.getLogger(__name__)
 
@@ -34,6 +40,7 @@ def launch(
     disk_gb: int = 64,
     mem_gb: int = 2,
     auto_clean: bool = False,
+    prepare_instance: "Callable[[Executor], None] | None" = None,
 ) -> MultipassInstance:
     """Create, start, and configure instance.
 
@@ -47,6 +54,8 @@ def launch(
     :param disk_gb: Disk allocation in gigabytes.
     :param mem_gb: Memory allocation in gigabytes.
     :param auto_clean: Automatically clean instance, if incompatible.
+    :param prepare_instance: A callback to perform early instance configuration
+        before the base image setup.
 
     :returns: Multipass instance.
 
@@ -79,5 +88,9 @@ def launch(
         mem_gb=mem_gb,
         image=image_name,
     )
+
+    if prepare_instance:
+        prepare_instance(instance)
+
     base_configuration.setup(executor=instance)
     return instance