+++ craft_providers/provider.py
@@ -26,9 +26,14 @@ import logging
 import pathlib
 from abc import ABC, abstractmethod
 from collections.abc import Generator
+from typing import TYPE_CHECKING
 
 from .base import Base
-from .executor import Executor
+
+if TYPE_CHECKING:
+    from collections.abc import Callable
+
+    from craft_providers import Executor
 
 logger = logging.getLogger(__name__)
 
@@ -79,7 +84,7 @@ class Provider(ABC):
         """
 
     @abstractmethod
-    def create_environment(self, *, instance_name: str) -> Executor:
+    def create_environment(self, *, instance_name: str) -> "Executor":
         """Create a bare environment for specified base.
 
         No initializing, launching, or cleaning up of the environment occurs.
@@ -98,7 +103,9 @@ class Provider(ABC):
         instance_name: str,
         allow_unstable: bool = False,
         shutdown_delay_mins: int | None = None,
-    ) -> Generator[Executor, None, None]:
+        use_base_instance: bool = True,
+        prepare_instance: "Callable[[Executor], None] | None" = None,
+    ) -> Generator["Executor", None, None]:
         """Configure and launch environment for specified base.
 
         When this method loses context, all directories are unmounted and the
@@ -112,4 +119,7 @@ class Provider(ABC):
         :param allow_unstable: If true, allow unstable images to be launched.
         :param shutdown_delay_mins: Minutes by which to delay shutdown when exiting
             the instance.
+        :param use_base_instance: Enable base instances if supported by the provider.
+        :param prepare_instance: A callback to perform early instance configuration
+            before the base image setup.
         """