+++ tests/unit/lxd/test_lxd_provider.py
@@ -18,6 +18,7 @@ from datetime import timedelta
 from unittest.mock import MagicMock, call
 
 import pytest
+from craft_providers import Executor
 from craft_providers.bases import ubuntu
 from craft_providers.errors import BaseConfigurationError
 from craft_providers.lxd import LXDError, LXDProvider, LXDUnstableImageError
@@ -171,6 +172,57 @@ def test_launched_environment(
                 project="default",
                 remote="local",
                 expiration=expiration,
+                prepare_instance=None,
+            ),
+        ]
+
+        mock_launch.reset_mock()
+
+    assert mock_launch.mock_calls == [
+        call().unmount_all(),
+        call().stop(delay_mins=None),
+    ]
+
+
+def test_launched_environment_prepare_instance(
+    mock_buildd_base_configuration,
+    mock_get_remote_image,
+    mock_remote_image,
+    mock_launch,
+    mock_lxc,
+    tmp_path,
+):
+    provider = LXDProvider(lxc=mock_lxc)
+    expiration = timedelta(days=90)
+
+    def _prepare_instance(instance: Executor) -> None:
+        pass
+
+    with provider.launched_environment(
+        project_name="test-project",
+        project_path=tmp_path,
+        base_configuration=mock_buildd_base_configuration,
+        instance_name="test-instance-name",
+        prepare_instance=_prepare_instance,
+    ) as instance:
+        assert instance is not None
+        mock_get_remote_image.assert_called_once_with(mock_buildd_base_configuration)
+        mock_remote_image.add_remote.assert_called_once_with(lxc=mock_lxc)
+        assert mock_launch.mock_calls == [
+            call(
+                name="test-instance-name",
+                base_configuration=mock_buildd_base_configuration,
+                image_name="test-image-name",
+                image_remote="test-remote-name",
+                auto_clean=True,
+                auto_create_project=True,
+                map_user_uid=True,
+                uid=tmp_path.stat().st_uid,
+                use_base_instance=True,
+                project="default",
+                remote="local",
+                expiration=expiration,
+                prepare_instance=_prepare_instance,
             ),
         ]